+++ src/form/Attributes/common/BaseInput/EditablePreview/index.js
@@ -1,0 +1,28 @@
+import React from 'react';
+
+import {EditablePreview as BaseEditablePreview} from '../../Base';
+import {useBaseInput} from '../common/hooks';
+import FormInput from '../common/Input';
+
+const NodeValueRender = ({node, nodeDef, keyboardType = 'default'}) => {
+  const {debouncedUpdate, handleSubmitEnter} = useBaseInput({
+    node,
+    nodeDef,
+  });
+
+  return (
+    <FormInput
+      debouncedUpdate={debouncedUpdate}
+      handleSubmitEnter={handleSubmitEnter}
+      node={node}
+      nodeDef={nodeDef}
+      keyboardType={keyboardType}
+    />
+  );
+};
+
+const EditablePreview = ({nodeDef}) => (
+  <BaseEditablePreview nodeDef={nodeDef} NodeValueRender={NodeValueRender} />
+);
+
+export default EditablePreview;
