--- src/form/Attributes/common/BaseInput/Form/styles.js
@@ -1,5 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({});
-
-export default styles;
