+++ src/form/Attributes/common/BaseInput/index.js
@@ -1,2 +1,3 @@
 export {default as Preview} from './Preview';
+export {default as EditablePreview} from './EditablePreview';
 export {default as Form} from './Form';
